LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-container-expected.txt


FAIL Style rules within @container are scoped assert_equals: expected "1" but got "auto"
